Toggle focus level instead of moving
Previously Esc and Alt+Enter moved focus level a level
up, or a level down, and hitting e.g. Esc repeatedly
kept the focus level, but moved to the previous item
on the same level. Now, Esc is simply an alternative
binding to Alt+Enter and Alt+Enter toggles the focus
level: hitting it repeatedly continues toggling the
level.