Make PageUp/PageDown based on non-relative pages
Instead of moving cursor to top/bottom of the screen,
the cursor is now moved to top/bottom of the actual
"physical page".

In physical pages the page offset is always same so that
the same number of PageDown/PageUp presses leads to the
same location.

Not sure if this will be a good idea -- may change this
later...