Change default font to monospaced
Previously the variable width font was there only for
testing purposes. Using a variable width font with standard
command line tools does not result in correct indentation.

If all output that has columns would use a single '\t'
as a field separator we could automatically adjust the columns
so that they would match no matter if we'd use variable width
or fixed width font.

In theory we could add a FS variable (field separator) as
we have it in sh/awk and adjust the output accordingly but
generally using a monospaced font is easier :)