Remove black gaps between buffers
Due to rounding error there was black gaps between the buffers,
and because the default color was different from the actual
color.

There is still an issue with gaps, but that is because there
is not yet "size increment" support in the layout system i.e.
we should give more space only if the amount is large enough
to satisfy a certain size e.g. so that a new row could
completely fit to the given space.