Consolidate drawing operations
This is mainly for performance and avoiding redrawing, especially
the kind which might cause flicker. Basically the goal is to
combine geometry changes and drawing operations to one change,
or one drawing operation per cycle. Not perfect yet.

Interesting change in this set is the fact editor size is not
made smaller until the command has finished, or until focus is
moved: this removes the majority of unnecessary redrawing by
minimizing global resizing.

Also added some compile-time options to config.h that can be
experimented with, until finding a combination that works well,
everywhere.